partially revert 06a18733329a67c0954d87c69ad0147fdfdce6ff - put back dockstring - compute mpi n_chains_per_rank for MPI in the base class as it was originally
